restructure & fixing #215
