+++ shared/src/commonMain/kotlin/post/RESTfulPostManager.kt
@@ -2,13 +2,11 @@ package post
 
 import account.AppUser
 import account.utilities.UUID
-import event.TrendWaveState
 import io.ktor.client.HttpClient
 import io.ktor.client.request.get
 import io.ktor.client.request.post
 import io.ktor.client.statement.bodyAsText
-import utilities.CommonLogger
-import utilities.DateUtil
+import utilities.textutils.DateUtil
 
 class RESTfulPostManager{
