+++ shared/src/commonMain/kotlin/utilities/textutils/DateUtil.kt
@@ -1,6 +1,7 @@
-package utilities
+package utilities.textutils
 
 import io.ktor.util.date.GMTDate
+import utilities.MonthCalc
 
 class DateUtil {
