+++ shared/src/commonMain/kotlin/utilities/textutils/closeKeyboard.kt
@@ -1,0 +1,5 @@
+package utilities.textutils
+
+import androidx.compose.ui.ExperimentalComposeUiApi
+
+expect fun closeKeyboard()
