+++ shared/src/iosMain/kotlin/utilities/textutils/closeKeyboard.kt
@@ -1,0 +1,9 @@
+package utilities.textutils
+
+import platform.UIKit.UIApplication
+import platform.UIKit.endEditing
+
+actual fun closeKeyboard() {
+    val keyWindow = UIApplication.sharedApplication.keyWindow
+    keyWindow?.endEditing(true)
+}
